Removing unused files and metrics references
