+++ EventHubs/src/EventHubs_Logs/function.json
