+++ EventHubs/src/lib/mainindex.js
@@ -4,7 +4,6 @@
 
 var Transformer = require('./datatransformer.js').Transformer;
 var SumoClient = require('./sumoclient.js').SumoClient;
-var SumoMetricClient = require('./sumometricclient.js').SumoMetricClient;
 var sumoUtils = require('./sumoutils.js');
 
 module.exports = {
@@ -12,6 +11,5 @@ module.exports = {
     "p_wait" : sumoUtils.p_wait,
     "p_retryTillTimeout" : sumoUtils.p_retryTillTimeout,
     SumoClient:SumoClient,
-    SumoMetricClient:SumoMetricClient,
     Transformer:Transformer
 }
