+++ EventHubs/src/package.json
@@ -7,7 +7,7 @@
   "devDependencies": {},
   "scripts": {
     "test": "",
-    "build": "cp index.js ../target/logs_build/EventHubs_Logs/ && cp index_metrics.js ../target/metrics_build/EventHubs_Metrics/index.js && cp ../../sumo-function-utils/lib/*.js ../target/logs_build/EventHubs_Logs/ && cp ../../sumo-function-utils/lib/*.js ../target/metrics_build/EventHubs_Metrics/"
+    "build": ""
   },
   "author": "Himanshu Pal",
   "license": "Apache-2.0"
