+++ README.md
@@ -8,7 +8,7 @@ Following integrations are present. For more info look at their respective ReadM
 
 | FunctionName | Description | Collection Use Cases | Setup Documentation
 | -------------| ----------- | -------------- | ------------------- |
-|[Sumo Logic Azure Event Hub Integration](EventHubs)| This [solution](https://help.sumologic.com/Send-Data/Collect-from-Other-Data-Sources/Azure_Monitoring) creates a data pipeline for collecting logs/metrics from Eventhub.It includes separate ARM Templates for Metrics and Logs.| [Azure Audit App](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure-Audit/Azure-Audit-App-Dashboards) [Azure Active Directory App](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure_Active_Directory/Install_the_Azure_Active_Directory_App_and_View_the_Dashboards) [Azure SQL App](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure_SQL/Install_the_Azure_SQL_App_and_View_the_Dashboards)| [Docs for Logs](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure-Audit/02Collect-Logs-for-Azure-Audit-from-Event-Hub) <br/> [Docs for Metrics](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure-Audit/02Collect-Logs-for-Azure-Audit-from-Event-Hub)|
+|[Sumo Logic Azure Event Hub Integration](EventHubs)| This [solution](https://help.sumologic.com/Send-Data/Collect-from-Other-Data-Sources/Azure_Monitoring) creates a data pipeline for collecting logs from Eventhub.It includes separate ARM Templates for Metrics and Logs.| [Azure Audit App](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure-Audit/Azure-Audit-App-Dashboards) [Azure Active Directory App](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure_Active_Directory/Install_the_Azure_Active_Directory_App_and_View_the_Dashboards) [Azure SQL App](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure_SQL/Install_the_Azure_SQL_App_and_View_the_Dashboards)| [Docs for Logs](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure-Audit/02Collect-Logs-for-Azure-Audit-from-Event-Hub) <br/> [Docs for Metrics](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure-Audit/02Collect-Logs-for-Azure-Audit-from-Event-Hub)|
 |[Sumo Logic Azure Blob Storage Integration](BlockBlobReader) | This [solution](https://help.sumologic.com/Send-Data/Collect-from-Other-Data-Sources/Azure_Blob_Storage) event-based pipeline for shipping monitoring data from Azure Blob Storage to an HTTP source on Sumo Logic.| [Azure Web Apps](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure-Web-Apps/Azure-Web-Apps-Dashboards) | [Docs for Logs](https://help.sumologic.com/Send-Data/Applications-and-Other-Data-Sources/Azure-Web-Apps/01Collect-Logs-for-Azure-Web-Apps) |
