--- sumo-function-utils/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "sumo-function-utils",
-  "version": "0.0.1",
-  "description": "This is an utility package for Sumo Logic Inc.",
-  "license": "Apache",
-  "files": [
-    "./lib"
-  ],
-  "main": "./lib/mainindex",
-  "devDependencies": {
-    "chai": "*",
-    "mocha": "*",
-    "https": "*",
-    "zlib": "*",
-    "url": "*"
-  }
-}
